In-SMEM, iterative/by-slice node construction working! On GTX 670, it is faster than the leaf building, so leaving for now. The idea is to iterate, each time building part (a slice) of the tree using only shared memory for flags etc.
Could do with some tidying. For reference, the main problem was in
recognizing nodes that would become 'base nodes', inputs for starting
the tree-climb in the next iteration. These base nodes could be:
a) the *only* child to write to their parent;
b) of sufficient size to become base nodes and written to by threads *in
the same block*; and
c) of sufficient size but written to by threads in *different blocks*.
Differentiating between b) and c) seemed non-trivial. In actuality, it can be
computed by post-processing each tree-climb iteration, provided the
post-processing can compute the size -- in terms of *base nodes* spanned --
of each node.
This required modification to the node slice building:
a) it outputs nothing to the work queue (a post-processing kernel does
that); and
b) until a node has been 'finalized' -- that is, had *both* its children
written -- its .z/.w values refer to the base-later nodes it spance, rather
compacted rather than the leaves.