Do not safe-size Triangle AABB's unless in DEBUG mode.
Bumping out the bounds of the AABB when triangles are zero-sized has a
minor impact on performance results, but is needed only to avoid failing
an assertion check when GRACE_DEBUG is defined.

It seems slightly fairer to compare against the SPH tree-build times
when the Triangle AABBs function does not have this unnecessary overhead.