More advanced triangle shading.
Per-light triangle shading is precomputed (and hence inexact, but in
general triangles are much smaller than the triangle-light distance, and
hence the error is small).
Shadow rays are cast to check whether an area should be illuminated.

Runs, but unclear if it works correctly. Adding a surface below the
model would make shadows obvious.