Add pinhole camera model rays.
Similar to a combination of gluPerspective and gluLookAt; that is,
a combination of perspective and view matrices. The position of the
camera, its (vertical) field-of-view, its orientation, and the point at
the centre of view are all required. Pixel resolution must also be
provided, and defines the aspect ratio; square pixels are always assumed
to simplify the API.

There is no far clipping-plane equivalent, but rather all rays have the
same length, and hence terminate on a spherical surface. This is for
simplicity in the implementation.