Orthographic projection and perspective projection APIs consistent.
Underlying algorithm is similar, except for the former points on the
image plane become the ray origins, and in the latter they are used to
compute ray directions.

Using a camera model for the orthographic projection is also more
consistent with the expected use-case; plane_parallel_random_rays still
maintains the interface previously used for orthographic projections for
that same reason.

A small amount of refactoring and tidying has also been done, and some
missing includes added.