Add CUDA RNG-state managing class and benchmark.
Future use-case is that it be a member of other classes which need
sources of randomness, e.g. ray-generating classes.

To save on both memory and initialization time, unless otherwise specified
the number of generated states is equal to the maximum number of threads
which may be simultaneously in-flight on the current device. The RNG
states are intended to be used in a persistent-threads model.