Fix bug in PRNG benchmark.
If the output of the RNG (e.g. curand_normal) call is not used, then it
appears that the compiler can optimize away some, or all, of the code
except the actual state transition. This gives misleading results,
particularly for Philox where the state transition is essentially an
increment operation!

Now, the output of the RNG is always used. Further, curand_normal is
called, since this is the distribution we actually care about; and three
numbers are generated per 'N', so the benchmark corresponds more closely
to generating 'N' rays.

The output table formatting has also been improved to better handle
different values of BITMASK (and hence the displayed probabilities).