Only one PRNG alloc. Can update PRNG seed and re-initialize.
Since the target device and number of states are fixed at initialization,
allocation of states need only be performed once.

The most common use-case for init_states() will likely be after an
update to the seed value. Hence it makes sense for this to be possible
with a single method call.

Also tidied the procedure for swapping to the internal device and back
to whatever the set device was on entry to the function.