PRNG init and gen times corrected.
The total time, not the average time, was being output. This has been
corrected. The formatting of the output has also been updated/fixed to
display correctly for all (most?) typical output values.

A new seed is provided on each call to init_states. This seems to have
no effect on the performance under CUDA 8.0, but may help avoid any
future compiler optimization that would otherwise contract the n_iter
calls down to a single call(?)