Fix code broken due to rebase.
Largely consists of:
  - float3 -> grace::Vector<3, float>
  - float3 bot/top -> grace::AABB<float>
  - updating GRACE calls which use the above
in code which was added to 'master' after cleaner-API branched from
it. As a result, it is largely the triangle-related testers and
profilers which have been modified.

Another side effect is the removal of float3-based vector math
operations, in favour of using the various grace::Vector math functions.

This also resulted in the addition of compound assignment operators for
the grace::Vector class. Note that they have been implemented as free
functions, not member functions. This is because we do not actually
define the generic Vector<Dims, T> class; attempting to define a generic
compound assignment operator is therefore a compiler error. We would
need to write each compound assignment op. definition for each
specialization we define, i.e. for both Vector<3, T> and Vector<4, T>.
This is pretty tedious, given that the member function bodies would
contain _identical_ code. The solution, then, is just to define said
operators as free functions. This will automatically scale if we declare
some new specialization, e.g. Vector<2, T>, in the future. The
implementation is likely sub-optimal now, since we construct a temporary
vector and then assign it to the Vector we are 'in-place' modifying. But
hopefully the compiler will be able to optimize most of that out anyway.
And/or it's just not a significant overhead.

There are a few other fixes of code which apparently got mangled in the
rebase.

All test compile, and all but tree_traversal and sphere_intersection
have been run (with GRACE_DEBUG defined at compile-time) to ensure they
exit with PASSED, or print sensible results, or produce a correct bitmap
image, as appropriate.