Add equivalence and noninferiority tests for isotropy.
The bulk of the work is in the newly-added .py files. They rely on the
existence of content in the ray_dumps folder, however, which is
generated by the new ray_dump.cu.

The readme has been updated to i) include a desription of Ripley's
K-function; and ii) describe the purpose of the various .cu and .py
files, and in particular point out that the .cu files do not rigorously
confirm isotropy!

The csv/, plots/ and ray_dumps/ folders are part of the repository to
avoid any bugs where a script or executable may fail to run because the
directory it needs hasn't been created. To avoid clutter in `git status`
after the tests/ray dumps have been run, the folders also have
.gitignore files to ignore their expected content. (This serves
double-duty, as it is not possible to track an empty folder in git.)