Update all profilers and testers for new CudaBvh.
MAJOR traversal performance regression. Kernels that would take < 100 ms
now take several seconds (GTX 670); most of the traversal profilers in
fact time out. nvvp is of little use --- it looks like basically all
kernels time-out when being profiled, even if they run stand-alone.

At least in isolation, it does not seem to be due to:
  - private: in CudaBvhNode/Leaf
  - Use of TexRefIter for node/leaf access.
  - Use of Vector3f in place of float3 for origin and invd in trace
    kernel.
  - Use of Vector4f for AABBs in CudaBvhNode.
  - Reading of entire CudaBvhNode 'at once'.
  - Use of Vector4i/f in CudaBvhNode.
  - Alignment of CudaBvhNode.

Which covers most of the differences between this implementation and the
old one, using manual textures and vectors of float4/int4 types...

One point of note, the number of registers for the trace kernel in
integrate_gadget has increased from 38 in a vaguely-current master to 46
here. Not sure that register spilling could have such a disasterous
impact on performance, but it is a concern in its own right.