Fix bug in AABB creation. Traversal performance now as before.
ALBVH builder assumes that AABBs start in an invalid state, so that
aabb_union(newly_created_aabb, other_aabb) == other_aabb. This is now
the case. For safety, AABBs are invalidated at creation.
bbox.invalidate() is too easy to forget.

As per issue #67, this was also best implemented through a
grace::numeric_limits struct which mimics std::numeric_limits but also
works on the device. grace::numeric_limits provides only the most
commonly-used parts of std::numeric_limits, and doesn't define
CUDA-compatible version of those which make no sense (e.g. ::infinity
for integral types). It also only specialized over int, unsigned int,
float and double. This is all mostly for time-saving on my part.
Certainly, at least [unsigned] long [long] should probably also be
specialized.

Also noticed some weirdness with _ISOC11_SOURCE, in that it seems to be
defined under normal compilation. Since it's actually for C code, not
C++ code, I am removing it.

This in turn revealed an issue with posix_memalign, in that it requires
the alignment argument to be a multiple of sizeof(void*), which on all
tested systems is 8, and hence the alignment test with alignment == 4
fails. This is rectified by increasing the alignment where it's less
than sizeof(void*), which is always okay because i) valid alignments are
always powers of two; and ii) an increased alignment will still fulfill
the originally-requested alignment.

aligned_alloc now checks for non-zero, power-of-two alignment for good
measure.