Remove raydata.h.
Hopefully the current many-functors system for customizing the traversal
will soon be gone. But for now, these RayData structs have been moved to
the SPH-specific traversal header, since that's the only place they are
used. Also moved into detail::, because they should not be considered
part of the public API.